chore: drop cargo-release from devshell after simit cutover
Phase 06 of the simit-integration plan. simit (added in fe195bf) now owns the
release path via cargo xtask release (8a3ff40); the in-CI smoke job (37e300a)
exercises the dry-run on every push. The cargo-release fallback in the devShell
is no longer needed.

Note: Phase 06's strict precondition was "at least one real release cut via
simit". No tag has been cut yet; the CI smoke job is the proxy. Revert is one
commit away if the simit path fails on a real release.

+++ crates/modde-games/tests/high_impact_games_tests.rs
@@ -125,7 +125,7 @@ fn stardew_scanner_discovers_smapi_mods() {
 
     assert_eq!(mods.len(), 1);
     assert_eq!(mods[0].mod_id, "smapi/BetterRanching");
-    assert!((mods[0].confidence - 0.98).abs() < f32::EPSILON);
+    assert!((mods[0].confidence - 0.98).abs() < f64::EPSILON);
 }
 
 #[test]
